NNA: adla driver android porting 1.7.1 [1/2]
PD#SWPL-203459

Problem:
adla android porting 1.7.1

Solution:
adla android porting 1.7.1
adlak driver 1.4.0.6.5 -> 1.7.1.0.0

Verify:
s5, s6, T7c, T3x

+++ adla/kmd/drv/common/mm/adlak_mm_cma.c
@@ -549,7 +549,7 @@ static struct adlak_mem_handle *adlak_cma_attach(struct adlak_context_smmu_attr 
     if (!ret) {
         pbuf_req->errcode = 0;
     } else {
-        pbuf_req->errcode = ret;
+        pbuf_req->errcode = (int32_t)ret;
         adlak_mem_uid_free(mm_info->uid);
         adlak_os_free(mm_info);
         mm_info = (void *)NULL;
